lambda/requirements.txt
ask-sdk-core
ask-sdk-model
boto3
google-generativeai
python-dotenv